+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `laravel-query-filters` will be documented in this file
 
+## 0.2.0 - 2021-12-14
+
+- Support Laravel 7.x, 8.x
+- Rename trait `Sortable` to `SortableQueryFilter`
+- Rename trait `Searchalbe` to `SearchableQueryFilter`
+- Remove travis and styleci
+- Remove `make:filter` command
+
 ## 0.1.0 - 2019-09-20
 
 - Support Laravel 6.x